+++ src/Symfony/Component/Mime/Part/TextPart.php
@@ -97,6 +97,14 @@ class TextPart extends AbstractPart
     }
 
     /**
+     * @return ?string null or one of attachment, inline, or form-data
+     */
+    public function getDisposition(): ?string
+    {
+        return $this->disposition;
+    }
+
+    /**
      * Sets the name of the file (used by FormDataPart).
      *
      * @return $this