+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Support detection of related parts if `Content-Id` is used instead of the name
+ * Add `TextPart::getDispostion()`
 
 6.2
 ---